refactor : 계층형 패키지 구조 변경
