+++ src/test/java/car/CarServiceTest.java
@@ -1,7 +1,7 @@
 package car;
 
 import car.domain.Car;
-import car.domain.CarService;
+import car.service.CarService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
